third_party/blink/web_tests/svg/parser/whitespace-length-invalid-3-expected.txt

CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "-.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, ".-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " - ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " + ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " +%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " -.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, " .-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   - ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   + ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   +%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   -.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "   .-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t - ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t + ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t +%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t -.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\r\n\t .-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-em".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-ex".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-px".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-in".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-cm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-mm".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pt".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pc".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-%".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-. ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.- ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-em ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-ex ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-px ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-in ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-cm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-mm ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pt ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pc ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-% ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-em   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-ex   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-px   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-in   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-cm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-mm   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pt   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pc   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-%   ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-em\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-ex\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-px\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-in\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-cm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-mm\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pt\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pc\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-%\r\n\t ".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f+%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f-.%\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-em\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-ex\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-px\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-in\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-cm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-mm\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pt\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-pc\f".
CONSOLE ERROR: Error: <svg> attribute x: Expected length, "\f.-%\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
